requirements.txt
Flask
pandas
pdf2image
pytesseract
Pillow
spacy
tqdm